+++ src/components/link/LocalizedLink.jsx
@@ -216,8 +216,6 @@ const LocalizedLink = (props) => {
   } else {
     console.log('Creating Contentful link to', props.to);
 
-    console.log(data);
-
     const node = data.allContentfulEntry.edges.filter(
       ({ node }) => node.contentful_id === props.to && node.node_locale === props.locale,
     )[0].node;