+++ src/templates/category.jsx
@@ -120,7 +120,7 @@ const Page = (props) => {
                 </Link>
               </H2>
               <PostMeta>
-                <time datetime={post.datetime}>{post.date}</time>
+                <time dateTime={post.datetime}>{post.date}</time>
                 <PostAuthor>
                   <AuthorName author={post.author} locale={props.pageContext.locale} />
                 </PostAuthor>